Minor fixes and tweaks
- Using char array instead of string for "WAVE" checking
- Updated README
- Realized 'pause' was broken so fixed it
- Changed buffer to 16-bit from 8, changed playback interrupt to handle
- Reverts pwm mode if sample rate too high

+++ TMRpcm/TMRpcm.h
@@ -17,14 +17,13 @@ class TMRpcm
 	boolean pwmMode;
 	boolean isPlaying();
 	boolean rfPlaying;
-	unsigned long SAMPLE_RATE;
+	unsigned int SAMPLE_RATE;
 
 
  private:
 	int lastSpeakPin;
 	void setPin();
 };
-
